FEAT: added authorisation, controller checked and finished

+++ MeFit.API/Controllers/GoalsController.cs
@@ -10,6 +10,7 @@ using MeFit.DAL.Models.Domain;
 using MeFit.DAL.Models.DTOs.Goal;
 using AutoMapper;
 using System.Net.Mime;
+using Microsoft.AspNetCore.Authorization;
 
 namespace MeFit.API.Controllers
 {
@@ -25,147 +26,159 @@ namespace MeFit.API.Controllers
             _mapper = mapper;
             _context = context;
         }
+
+        // GET: api/goals
         /// <summary>
-        /// Gets list of all goals with workouts in it
+        /// Gets all goals owned by the user
         /// </summary>
-        /// <returns>List of goals</returns>
+        /// <returns>List of all goals</returns>
         /// <response code="200">Returns all goals</response>
         /// <response code="204">No goals found</response>
-        // GET: api/Goals
+        /// <response code="401">Not authorized</response>
         [HttpGet("goals")]
+        [Authorize]
+        [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<ActionResult<IEnumerable<GoalReadDTO>>> GetGoals()
         {
-            var goalReadDTO = _mapper.Map<List<GoalReadDTO>>(await _context.Goals.Include(w => w.Workouts).ToListAsync());
+            var usernameToken = TakeUserNameFromToken();
+            var userId = TakeIdFromUser(usernameToken).Result;
 
-            if (goalReadDTO.Count == 0)
+            var goals = _mapper.Map<List<GoalReadDTO>>(await _context.Goals.Include(g => g.Workouts).Include(g => g.Profile).ThenInclude(p => p.User).Where(g => g.Profile.User.Id == userId).ToListAsync());
+
+            if (goals.Count == 0)
             {
                 return NoContent();
             }
 
-            return Ok(goalReadDTO);
+            return Ok(goals);
         }
 
-        // GET: api/Goals/5
+        // GET: api/goal/5
         /// <summary>
         /// Gets goal by ID
         /// </summary>
         /// <param name="id">ID of a goal</param>
         /// <returns>Goal</returns>
         /// <response code="200">Returns a goal</response>
+        /// <response code="401">Not authorized</response>
+        /// <response code="403">Not allowed</response>
         /// <response code="404">No goal found</response>
         [HttpGet("goal/{id}")]
+        [Authorize]
         [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<ActionResult<GoalWorkoutsDTO>> GetGoalById([FromRoute] int id)
         {
-            var goalReadDTO = _mapper.Map<GoalWorkoutsDTO>( await _context.Goals.Include(g => g.Workouts).FirstOrDefaultAsync(g => g.Id == id));
-
-            if (goalReadDTO == null)
+            var usernameToken = TakeUserNameFromToken();
+            var userId = TakeIdFromUser(usernameToken).Result;
+            //check if goal exists
+            var goal = _mapper.Map<GoalWorkoutsDTO>(await _context.Goals.Include(g => g.Workouts).FirstOrDefaultAsync(g => g.Id == id));
+            if (goal == null)
             {
                 return NotFound();
             }
+            //check if the user is owner of the goal
+            var profile = await _context.Profiles.Include(p => p.User).Include(p => p.Goals).Where(p => p.User.Id == userId).FirstOrDefaultAsync();
+            if (!profile.Goals.Any(g => g.Id == goal.Id))
+            {
+                return Forbid();
+            }
 
-            return Ok(goalReadDTO);
+            return Ok(goal);
         }
 
+        // PATCH: api/goal/5/SetAchieved
         /// <summary>
         /// Updates the corresponding goal to achieved
         /// </summary>
         /// <param name="id">ID of a goal</param>       
-        /// <param name="profId">ID of a profile</param>
         /// <response code="204">Successfully changed goal to achieved</response>
         /// <response code="400">Bad request</response>
+        /// <response code="403">Not allowed</response>
         /// <response code="404">No goal/profile found</response>
-        // PATCH: api/Goals/5/SetAchieved
+        /// <response code="500">Internal Server Error</response>
         [HttpPatch("goal/{id}/SetAchieved")]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
-        public async Task<IActionResult> UpdateGoalAchieved([FromBody] int profId, [FromRoute] int id)
+        public async Task<IActionResult> UpdateGoalAchieved([FromRoute] int id)
         {
+            var usernameToken = TakeUserNameFromToken();
+            var userId = TakeIdFromUser(usernameToken).Result;
+            //check if goal exists
             var goal = await _context.Goals.FindAsync(id);
-            var profile = await _context.Profiles.Include(p => p.Goals).FirstOrDefaultAsync(pi => pi.Id == profId);
-
-            if (profile == null)
+            if (goal == null)
             {
                 return NotFound();
             }
-            else if (goal == null)
+            //check if the user is owner of the goal
+            var profile = await _context.Profiles.Include(p => p.User).Include(p => p.Goals).Where(p => p.User.Id == userId).FirstOrDefaultAsync();
+            if (!profile.Goals.Any(g => g.Id == goal.Id))
             {
-                return NotFound();
-            }
-            else if (!profile.Goals.Any(p => p.Id == id))
-            {
-                return BadRequest();
+                return Forbid();
             }
 
             profile.Goals.FirstOrDefault(g => g.Id == id).Achieved = true;
-            await _context.SaveChangesAsync();
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
 
             return NoContent();
         }
 
-        // POST: api/Goals
+        // POST: api/goal
         /// <summary>
         /// Creates a goal
         /// </summary>
         /// <param name="newGoal">Goals info</param>
         /// <returns>A newly created goal</returns>
         /// <response code="201">Successfully created goal</response>
+        /// <response code="401">Not authorized</response>
+        /// <response code="500">Internal Server Error</response>
         [HttpPost("goal")]
+        [Authorize]
         [Produces(MediaTypeNames.Application.Json)]
         [ProducesResponseType(StatusCodes.Status201Created)]
         public async Task<ActionResult<GoalReadDTO>> PostExercise([FromBody] GoalCreateDTO newGoal)
         {
+            var usernameToken = TakeUserNameFromToken();
+            var userId = TakeIdFromUser(usernameToken).Result;
+            var profile = await _context.Profiles.Include(p => p.User).Where(p => p.User.Id == userId).FirstOrDefaultAsync();
+
             var domainGoal = _mapper.Map<Goal>(newGoal);
+            domainGoal.ProfileId = profile.Id;
             _context.Goals.Add(domainGoal);
-            await _context.SaveChangesAsync();
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
 
             return CreatedAtAction("GetExerciseById", new { id = domainGoal.Id }, _mapper.Map<GoalReadDTO>(domainGoal));
         }
 
-        /// <summary>
-        /// Adds list of goals in profile
-        /// </summary>
-        /// <param name="id">ProfileID</param>
-        /// <param name="goals">GoalsIDs to be added</param>
-        /// <response code="204">Successfully added goals to profile</response>
-        /// <response code="404">No profile found</response>
-        // POST: api/Goals/AddToProfile5
-        // Vily: little bit confused about this endpoint
-        [HttpPost("goals/AddToProfile{id}")]
-        [ProducesResponseType(StatusCodes.Status204NoContent)]
-        public async Task<IActionResult> CreateGoal([FromRoute] int id, [FromBody] List<int> goals)
-        {
-            var profile = await _context.Profiles.Include(p => p.Goals).FirstOrDefaultAsync(pi => pi.Id == id);
-
-            if (profile == null)
-            {
-                return NotFound();
-            }
-
-            foreach (var goalid in goals)
-            {
-                var tempGoal = await _context.Goals.FirstOrDefaultAsync(g => g.Id == goalid);
-                if (tempGoal != null)
-                {
-                    profile.Goals.Add(tempGoal);
-                }
-
-            }
-
-            await _context.SaveChangesAsync();
-
-            return NoContent();
-        }
-
+        // DELETE: api/goal/5
         /// <summary>
         /// Deletes a goal by ID
         /// </summary>
         /// <param name="id">ID of a goal</param>
         /// <response code="204">Successfully deleted goal</response>
+        /// <response code="401">Not authorized</response>
+        /// <response code="403">Not allowed</response>
         /// <response code="404">No goal found</response>
-        // DELETE: api/Goals/5
+        /// <response code="500">Internal Server Error</response>
         [HttpDelete("goal/{id}")]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status204NoContent)]
         public async Task<ActionResult> DeleteGoal(int id)
         {
@@ -174,15 +187,37 @@ namespace MeFit.API.Controllers
             {
                 return NotFound();
             }
-            _context.Goals.Remove(goal);
-            await _context.SaveChangesAsync();
+            var usernameToken = TakeUserNameFromToken();
+            var userId = TakeIdFromUser(usernameToken).Result;
+            //check if the user is owner of the goal
+            var profile = await _context.Profiles.Include(p => p.User).Include(p => p.Goals).Where(p => p.User.Id == userId).FirstOrDefaultAsync();
+            if (!profile.Goals.Any(g => g.Id == goal.Id))
+            {
+                return Forbid();
+            }
+
+            try
+            {
+                _context.Goals.Remove(goal);
+                await _context.SaveChangesAsync();
+            }
+            catch
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError);
+            }
 
             return NoContent();
         }
 
-        private bool GoalExists(int id)
+        private string TakeUserNameFromToken()
         {
-            return _context.Goals.Any(e => e.Id == id);
+            var username = User.Claims.FirstOrDefault(c => c.Type == "preferred_username");
+            return username.Value;
+        }
+        private async Task<int> TakeIdFromUser(string usernameFromToken)
+        {
+            var id = (await _context.Users.FirstOrDefaultAsync(x => x.Username == usernameFromToken)).Id;
+            return id;
         }
     }
 }
